requirements.txt
Flask
flask-login
Flask-JWT-Extended
Flask-SQLAlchemy==3.0.5
flask-cors
python-decouple
requests
celery
redis
pandas
openai
backoff
tiktoken
scipy
gunicorn
PyJWT
# psycopg2